Update patch set 2
Patch Set 2: Code-Review+2

>> Look, the #3 complaint I hear about Haiku after "it doesn't run on my hardware" and "the web browser doesn't load X website properly" is "it isn't secure / has no multiuser segmentation / etc." We need to address these concerns to get more adoption, and thus more momentum, funding, etc. etc. etc.

> This really is a question of principle. Do you take shortcuts or not? For me as a developer possibly targetting a platform, I want to know such things. If I realize that the people maintaining that platform will try hard to do changes properly, so that my software is not likely to break arbitrarily, this makes the platform attractive. If I have to fear that it's a moving target that someone will simply change underneath me, this is a big concern and possible reason not to invest in the platform at all. This then leads to the burden of porting things remaining in people at HaikuPorts and never spreading out to the actual projects. This IMHO is far worse for the momentum.

Come to think of it, *this* is what I see as the #3 item in the list. People complain that we don't have a stable release and they don't want to work on a moving target platform. This is what they mean. By taking such shortcuts in how we deploy changes, we reinforce that impression that Haiku is a moving target, and what works one day may not work the next day.

Security is of course also important and no one is questioning your changes here, just how they are deployed. I am well aware of the problems with not enforcing the protection on areas, but as mmlr mentionned, there are so many other ways to access the data anyway that I don't see how there is an emergency in plugging this particular hole. We have to do it, sure, but wether it becomes effective now or after beta2 and people had time to adjust their apps doesn't make much of a difference.

So yes, let's aim at "Microsoft level", no matter how long it takes.


Now, can we draft a plan that works best for this? It sure would be nice if old binaries do not plague us for years and that the backward compatibility cannot easily be used as a way to break the new check (by injecting a lib or add-on pretending to use the old API, I guess?).

Given that, I would say that using the executable API version might actually be enough. If the app developer has tested their application and it works, even with an old lib, then we can assume things are fine. If they tests their freshly built application with old libs and finds that this doesn't work, they can adjust dependencies to point at newer versions, or ask the library provider to fix the problem and add the flag. This sounds like a good compromise for me.

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+2