HaikuDepot: Display User's Usage Conditions
A HaikuDepotServer may have previously agreed to
user usage conditions.  If this is the case then
they are now able to view those conditions.  There
were some problems with date formatting in this
change and so the date formatting logic has moved
into the HaikuDepot source code temporarily until
issues with the BDateTime class can be resolved;
most likely as a separate piece of work.

Relates to 15209

Change-Id: Ic3e5413d9139f410d7f7e8b566d4c56352dd2778