Update patch set 1
Patch Set 1:

(8 comments)

Hi Stephan;  Thank you for the review.  I would be interested in your thoughts on my use of the "debugger" function.

> Is it too late to suggest to call it either "UsageConditions" 

Subject to opinion, but yes "UsageConditions" might have been good.  It's really everywhere now (incl. server, API, database etc...) so it would be better if it just stays as is or is changed across both projects + database in one piece of work later so that it is consistent.

> Why the change from username to nickname?

Sometime early on, the "username" of the user was called "nickname".  This is consistently in the server and API as "nickname" from that time.  It was a bit inconsistent in the desktop application so I'm trying to unify on "nickname" to avoid confusion.

> Why was the avatar functionality removed?

This is not being used at all and there is no plan medium-term to implement that.  It is best to remove the unused member variable to avoid confusion.  It can be re-introduced if/when time permits to implement a user avatar feature at some point in the future.

(patch 2 coming soon...!)

Patch-set: 1
Reviewer: Gerrit User 1000018 <1000018@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>