Update patch set 1
Patch Set 1:

I think the problem is still there for a picture that has a ratio somewhere between square and the shape of the frame (so slightly vertical rectangle, but not as much as the frame)?

I think what we need to do here is for example compute the picture aspect ratio (H/W, from srcRect in the code) and the frame aspect ratio (h/w, from "size" in the code). Then:

if (H/W > h/w): the picture is more vertical than the frame, we need to reduce w and add vertical letterboxing
if (H/W < h/w): the picture is more horizontal than the frame, we need to reduce h and add horizontal letterboxing
if the ratios are equal: do nothing, the picture already fits perfectly

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>