Add flag for Friulian language
It's the historical flag of the Friuli region [1], which according
to our Friulian language manager would be appropriate.

I'm not sure about the naming of the icon. "fur Friuli" might be
correct; we name the catalogs "fur.catkeys", for example. The language
codes ISO 639-2 and ISO 639-3 are "fur", too, so maybe "fur Friulian"
would be even more correct...(?)

I'm also not sure how we get from the IOM file for the flag to the icon
in e.g. the locale preferences. I'd need some input on that and will
make changes accordingly.

[1] https://en.wikipedia.org/wiki/Friuli