Update patch set 1
Patch Set 1:

> Patch Set 1:
> Using a country flag to represent a language is not appropriate. Do not do it. If you want a visual clue, come up with something else. Otherwise I'm going to add random country flags to all menu items and buttons in Haiku because "visual clues".

Really, you see adding a French flag to the "Rotate clockwise" menu item as the same thing as adding it to a French language option?

I do know that country and language don't map 1:1, but as it has been used for decades (probably longer) everywhere from the web to advertisments, and it serves the purpose to quickly find what you're looking for in a long list of localized language names, I think we can get away not being absolutely 100% pure here.

And we've already been doing it for a decade in Haiku.

Patch-set: 1
Reviewer: Gerrit User 1000004 <1000004@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>