Update patch set 1
Patch Set 1:

> Patch Set 1:
> my grandparents native Béarnese (https://en.wikipedia.org/wiki/B%C3%A9arnese_dialect) which my father also speaks natively although he rarely uses it nowadays.

Get him to translate Haiku and we'll add the Béarnese flag. :)

> Pretending that French is a France-only thing is not right.

It's just a representation of the (orgigin of the) language. This is only a matter for the user documentation, though. The Locale prefs show all countries' flags under the language.

> Pretending that it is the single language of France is also not right. So there is no reason they should share a flag.

Probably spoken by the majority of the country. When we get a localization of a region in a country, like Friulian in Friuli, we can add their flag.

Anyway, I know that you're formally right. I'm apparently not "woke" enough or lack the national streak to see it as such a big deal. Without those flags Haiku becomes a bit blander, and it's becoming more difficult for me to see that Русский mean Russian and Українська Ukrainian.

Patch-set: 1
Reviewer: Gerrit User 1000004 <1000004@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>