Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > my grandparents native Béarnese (https://en.wikipedia.org/wiki/B%C3%A9arnese_dialect) which my father also speaks natively although he rarely uses it nowadays.
> 
> Get him to translate Haiku and we'll add the Béarnese flag. :)

There is no flag, because it's not a country (and has not been since flags were invented).
My father is French, and speaks one of the languages of France.

On my mother side, my grandparents learnt French in school in Morocco after flewing from Spain during Franco ruling. Only much later they did move to France.

So I don't have to look very far (just my close relatives) to show you how a language isn't in any way tied to a country.

So one one side of my family, you can go back to the 1600s and still find people in France; none of which spoke French. And on the other side, you leave France at the second generation, yet French is still the language used. Which side should get to use the French flag, then?

I could dive into the political/social implications of this (that you can be considered "not French" for speaking another language, for example). And this is in a relatively peaceful and problem-free country. I can imagine how it can be much worse in other places. So yes, a flag in a menu is not taking active part in this, of course, yet, it turns my family language into a second-class one, which shouldn't be. I'm quite annoyed that my father chose to not transfer this language to me, and as a result a lot of my family culture (traditional songs, etc) may be getting lost (it's harder for me to remember lyrics to songs not in a language I can understand).

> > Pretending that French is a France-only thing is not right.
> 
> It's just a representation of the (orgigin of the) language. This is only a matter for the user documentation, though. The Locale prefs show all countries' flags under the language.

And there is a very good reason I designed the Locale preferences this way. I would have preferred a simpler look with less things to pick from, but it turns out that is not possible.

> 
> > Pretending that it is the single language of France is also not right. So there is no reason they should share a flag.
> 
> Probably spoken by the majority of the country. When we get a localization of a region in a country, like Friulian in Friuli, we can add their flag.
> 
> Anyway, I know that you're formally right. I'm apparently not "woke" enough or lack the national streak to see it as such a big deal. Without those flags Haiku becomes a bit blander, and it's becoming more difficult for me to see that Русский mean Russian and Українська Ukrainian.

If that's the only problem, I think last time we agreed (on a suggestion from KapiX) that showing the ISO language code would be fine. And in any case, why would you care about a language you don't understand anyway? Or if you do, why not use the opportunity to learn a little of their alphabet?

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>