Update patch set 1
Patch Set 1:

> I once more made the common connection between Béarnese and the region it's spoken, Béarn, which has a flag.

That flag was made up in recent days. It has a coat of arms, but it has been part of France since 1620 and that was a peaceful merge.

If you want to use flags from the 1620s, that's fine, but then the flag used for English is clearly not right, since it is a mashup of the flags for scotland, wales, and england. So, we should use this one: https://en.wikipedia.org/wiki/England

I can come up with more and more examples of this, let's imagine we are 20 years ago. Would you represent German using the West Germany or East Germany flag? Would you use the Czecho-Slovak flag to represent the Czech language? The Slovak language? Or both?

Situations like this happen all the time. Trying to use flags for languages is a political statement, in one direction or another. It may agree somewhat with the current situation in the EU, but it didn't 20 years ago, and may not agree again in a not-so-distant future, and certainly doesn't hold in other places.

> Anyway, since I doubt many follow this ticket that only adds an IOM file, but our conversation deals with much deeper changes, how about we take it to the dev ML?

It had already been discussed at https://dev.haiku-os.org/ticket/15311

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>