Update patch set 2
Patch Set 2:

The welcome page is already WebPositive start page, so it isn't very hard to find that way.
I think the more things we put on the desktop, the less likely it is that people will actually use them. At some point it just looks like a pre-cluttered desktop and you don't dig in the details of each icon, whereas if there is only one or two, you are more likely to identify them individually and not as "a bunch of random icons".

Patch-set: 2