Update patch set 7
Patch Set 7: Code-Review-1

Hi, the patch is now pretty clean and nicely written. However, I still think it is wrong to give a PackageView keyboard focus when it is selected. The PackageView exists for the purpose of graphically representing a package "item" in the FeaturedPackagesView. That it now needs to know about it's owner in order to learn about its siblings for the purpose of keyboard navigation, illustrates why the code is wrongly placed: The FeaturedPackagesView should be the one implementing keyboard navigation. It will (should) also be the one displaying keyboard focus, by drawing a focus border around the entire area (its frame). The code simply needs to be moved from PackagesView::KeyDown() to FeaturedPackagesView::KeyDown().
Keyboard focus should be received in FeaturedPackagesView when the user either clicks into the view, or when the user uses TAB to cycle focus into the view. Selecting a package item can be caused by other events than clicking into the view! I will never approve the patch as is when simply setting a package to selected state programmatically also has the side effect of shifting keyboard focus into the view. That is wrong. This solution focuses on the one test case that this patch is addressing, while breaking other situations. I understand that this may feel "frustrating" to the original contributor, but that is why we have code reviews in the first place. The system is not meant to approve patches without critique simply because they took a lot of work already. Comments such as these, as much as I understand them, make me feel frustrated myself. https://dev.haiku-os.org/ticket/11675#comment:10

Patch-set: 7
Reviewer: Gerrit User 1000024 <1000024@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1