Update patch set 5
Patch Set 5:

I have made the required changes and this seems to be working now.

Basically, it is not a good idea to try to add the attributes directly to the source file. A build that modifies its own sources is going to create problems. So, I added a rule that generate a copy of the quicktour file, and then add the icon to that. In order to keep things easier to follow, I grouped all the needed files in a single place (src/data/bin).

The rule is similar to the one used to add MIME types to the "tracker new templates"

Note for later: I find the use of a zip file for the Web+ bookmarks unconvenient. I think we could generate them from rdef files instead, or even straight from a Jamfile, so they are available in plaintext and easier to edit.

Patch-set: 5
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>