BListView: Update selection while holding mouse down
This feature was originally implemented in hrev50495.
This feature was removed (accidentally) when we switched to mouse up
in hrev52062.

In BeOS R5 as you scrolled through a single-selection list with the
mouse button held down the selected item appeared to change, but the
selection didn't actually update until you released the mouse button.

This commit copies the above behavior. hrev50495 changed the
selection as you scrolled, while  this commit the selection is
changed one time on mouse up instead. The behavior in this commit is
still different than BeOS R5 because the selection first changes when
you mouse down on the first item and then the selection potentially
changes again when you mouse up on the second item. In BeOS the
selection would change only one time on mouse up.

Fixes #15009

Change-Id: Icae02b8d37ed281390647504b4efa3d694ea522a