BListView: Update selection while holding mouse down
This feature was originally implemented in hrev50495.
This feature was removed (accidentally) when we switched to mouse up
in hrev52062.

In BeOS R5 as you scrolled through a single-selection list with the
mouse button held down the selected item appeared to change, but the
selection didn't actually update until you released the mouse button.

This commit copies the above behavior better than hrev50495 which
changed the selection as you scrolled, now the selection is changed
on mouse up instead. The behavior on Haiku is still slightly different
than BeOS R5 because on Haiku the selection changes when you moues down
on the first item and then the selection changes again when you mouse up
on the second item. In BeOS the selection would only change
one time on mouse up.

Fixes #15009