BListView: Update selection on MouseUp, not MouseDown
A version of this feature was originally implemented in hrev50495
that allowed you to scroll through a list of list items while the
mouse was held down updating the selection as you went.

This feature was removed when we switched to selecting on mouse up
in hrev52062.

This feature was never reimplimented when we switched back to
selecting on mouse down in hrev52121.

In BeOS R5 as you scrolled through a single-selection list with the
mouse button held down the selected item appeared to change, but the
selection didn't actually update until you released the mouse button.
The selection did not change on MouseDown, only on MouseUp. You could
click on one item then move your mouse off the first item to another
item and it would select the second item without ever selecting the
first item.

This commit duplicates the behavior so that Haiku behaves identically
to BeOS R5.

There is an exception to the rule that selection only updates on
MouseUp - when a drag is initiated on an unselected item then that
item is selected before initiating the drag. You may extend or alter
the selection by holding down shift or command while dragging. If the
clicked item was already part of a selection then the unaltered
selection is used for the drag instead.

hrev50495 was an insufficient copy of the above behavior as it
changed the selection as you scrolled, while in this commit the
selection appears to change as you scroll but in actuality the
selection is only actually changed one time on mouse up (or when
a drag is initiated on an unselected item).

Create private _DoSelection() method and call it either just before
initiating a drag or on MouseUp to perform a selection taking into
account modifier keys.

ScrollToSelection() in _Track() for autoscroll.

Eliminate MouseDownThread.

Fixes #15009 (and doesn't cause regression for #9190 #14264 #14289)

Change-Id: Icae02b8d37ed281390647504b4efa3d694ea522a