Tracker: fix fall back to bitmap icon
We can convert a bitmap icon to B_RGBA and scale it at the same
time if we pass in the right size. Don't error out, force icon
size to be either B_LARGE_ICON or B_MINI_ICON and then grab the
bitmap at that size and then convert and scale it.

Fixes #9554