Fix fallback for old repository config files (url vs. baseurl)
From what I understand, BRepositoryConfig::URL() is supposed to
be a unique identifier (uniquely identifying the repository across
mirrors), while BRepositoryConfig::BaseURL() specifies the URL
where the remote server (the repository mirror) can be reached.
The names for these identifiers seem strange to me. I would have
kept URL() and introduced a new getter Identifier() or ID() or
something. In addition, even for the new names, the meaning seems
to be reverse to intuition.
In any case, when the new BaseURL() field was introduced, the
repository config files (which are in driver-settings format)
have been updated to a new version. When they lack the key
"cfgversion", then they are version 1 files, otherwise whatever
cfgversion is specified. The code has a fallback for defining
BaseURL and URL from the old version of the files which have only
a "url" key. This commit tries to fix the fallback mechanism by
simply using the same string for url and baseurl.
This fixes software which tries to identify a repository by their
URL(), which was an empty string before this change. An example
is ticket #14927.