Fix fallback for old repository config files (identifier vs. baseurl)
When the new BaseURL() field was introduced, the
repository config files (which are in driver-settings format)
have been updated to a new version. When they lack the key
"cfgversion", then they are version 1 files, otherwise whatever
cfgversion is specified. The code has a fallback for defining
BaseURL and Identifier from the old version of the files which have only
a "url" key. This commit tries to fix the fallback mechanism by
simply using the same string for identifier and baseurl.
This fixes software which tries to identify a repository by their
Identifier(), which was an empty string before this change. An example
is ticket #14927.

Change-Id: Ib4b3857254b7b703858eff6815e2d6c54d69da3c