Update patch set 1
Patch Set 1:

Why is the shortcut for resizing not working as expected? Shouldn't we fix the shortcut based resizing rather than preventing it altogether?

The dragger is small and somewhat hard to reach, so I expect the use of the shortcut to be more convenient for people who know it.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>