Update patch set 2
Patch Set 2:

Hi Kacper, do I understand correctly, that with this API you can draw bitmaps that "wrap around"? If the bitmap is like this:

1234
5678
9abc

you could draw it like above with a "phase" of (0,0), but if you pass for example (1,0) as phase, then the bitmap would be drawn like this:

2341
6785
abc9

Correct?

Patch-set: 2
CC: Gerrit User 1000024 <1000024@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>