Deskbar: Improve auto-raise and auto-hide
The area to activate auto-raise was set to the prevent hiding area
for auto-hide, Instead use the window frame. Set the event mask to
B_POINTER_EVENTS when auto-raise or auto-hide is on and set the
event mask back to 0 when we turn off auto-raise and auto-hide.
Before this change you had to click a window twice to get it to
raise on top of a raised Deskbar, with this change you only have
to click once to make it rise above Deskbar.

The auto-hide show and hide bounds are unchanged.

There was another bug where the window was not hiding correctly in
auto-hide mode when you moused over the status bar in horizontal
mode. This was happening because the where parameter of
TBarView::MouseMoved and TBarView::MouseDown was yielding
coordinates relative to the status tray when the mouse was over
the status tray and not the current view meaning that
TBarView::Frame().Contains(where) would return false over the
status tray. Inspecting the where parameter showed that the x-
coordinate was reset back to 0 when you mouse over the status tray.
To fix this issue I pulled the screen_where field out of
CurrentMessage() instead since this yields the correct value.

The calendar window input focus has been fixed in auto-raise mode
so that you can click on calendar even when it is above Deskbar.

Don't hide Deskbar when Calendar is showing in auto-hide mode.

Fixes #8932 #14493