Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Ok, sounds good. Have you also tested drawing a tiled bitmap with an additional scaling transformation?

I'm not sure what you mean. Scaling as it works in DrawBitmap doesn't make sense in tiled case.

If you are asking about affine transforms of source bitmap in general: yes, I did experiment with that, and it works, but as noted in the ticket, it needs some refactoring to pass additional arguments.

For now I wanted to have a simple case done, and affine transforms can be added later. That is also why I introduced new APIs - it is easier to extend them, as opposed to introducing a temporary solution with DrawBitmap options parameter, that we would have to support since apps started using it. (My understanding is IK <-> app_server protocol is private.)

Patch-set: 2
Reviewer: Gerrit User 1000008 <1000008@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>