Update patch set 2
Patch Set 2:

I noticed that the new API doesn't allow to scale the bitmap, since only one rect is passed. But if setting an affine transformation on the view before using the API is possible, then everything is fine. IMHO it is important that if you have a sequence of drawing calls, including the new methods introduced in this commit, then the expected result of setting a transformation is that everything looks exactly the same, only transformed.

Patch-set: 2