Update patch set 2
Patch Set 2:

> Patch Set 2:
> I read through the ticket now. I understand that affine transformations are not yet correctly applied for these new drawing methods. It's of course OK to commit the patch regardless, we have tons of other functionality which doesn't work in all combinations.

Ok, I get it now. To make sure we are talking about the same thing, there are two types of affine transforms in this context:

1. src bitmap affine transform
2. dest rect affine transform

1. was tested, 2. was not. I don't see any reason why it wouldn't work though - if dest affine transforms worked fine for DrawBitmap, they should work the same for DrawTiledBitmap. As you can see in DrawBitmapGeneric scale is just set to 1.0, the rest is the same.

> I want to take a closer look at the changes in BitmapPainter, I haven't yet had the time to understand this more deeply.

It looks like a lot of changes, but it's meant to skip optimized painters when drawing in tiled mode. It would be cleaner with goto, but there is an allocation in the middle and gcc complains about it (rightfully so). Again, they can be changed/implemented later if needed.

Patch-set: 2
Reviewer: Gerrit User 1000008 <1000008@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>