Update patch set 1
Patch Set 1:

It already mounts the volume as read-only if replaying the log failed. I guess your intention is to run checkfs on the disk?

While that would work, you would need to have an extra disk if the boot disk is affected by this. Wouldn't it be nicer to force remount read/write instead?

Apart from this, the only thing that I would like to see changed is the code duplication that resets the log counter.

Patch-set: 1
Reviewer: Gerrit User 1000016 <1000016@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>