[RFC] Do not force deletion of SharedBufferList
The SharedBufferList is supposed to be reference counted, and stay
available as long as there are buffers allocated, since it manages them.

If the media roster deletes the area where it is allocated, there is no
point in that memory management.

Fix opensound/media_server crash on shutdown, as the opensound play
thread is still live when media_server destructs its MediaRoster.

I don't know if this is the proper fix: we could also make sure the
opensound thread is stopped before the roster, and in that case, remove
all the reference counting logic in SharedBufferList and have its
lifecycle managed entirely by BMediaRoster.