Add notice about locale-controlled 12/24-hour setting in Deskbar time preferences
To avoid confusion as to where exactly the setting to switch between 12-hour and 24-hour time display is located, this commit adds a notice in the Time preferences, the first logical place to look.
Potentially this could be replaced with an override checkbox, but in the interest of not breaking the current system, a notice should be enough.

Change-Id: Ifb935e6ad028c497268f18473fc01ef2c00f92cc