Update patch set 1
Patch Set 1:

> I'm going to leave this here: https://www.freelists.org/post/haiku-commits/haiku-hrev43984-srcappsdeskbar

Thanks for the reminder. Upon rethinking it, I'd say we made the right decision back in 2012.
The 12/24 hour setting should stay with the rest of formatting choices in the Locale prefs. A checkbox in Time|Clock could give the false impression that it's just a setting for the Clock of the Deskbar, not system-wide.

Thinking about this more, the whole Clock tab doesn't fit well into the Time prefs: All settings are limited to Deskbar's Clock replicant, they are not a system-setting as the rest of the Time prefs. I suppose it'd be best in it's own "Clock settings" window that would be invoked from the Clock replicant's context menu.

Anyway, to get on-topic again, I'd be in favour of a note where to set 12/24 hours mode. If we are agreed, I'd have a few suggestions for improved visuals for this patch.

Patch-set: 1