Update patch set 1
Patch Set 1:

So I've been looking at this a bit closer yesterday.
Here is what I found out (but it will end up being a more invasive change)

As I mentionned, there will always be people looking for time format settings in the Time preferences instead of Locale. Not just for the DeskBar settings, but also for the system-wide ones. So, we should find a way to have these closer together.

I don't like the idea of "notes" for this. It feels like doing only half of the work. We have noticed that there is a problem, but instead of fixing it, we are just telling the user to manually go elsewhere.

This is why I'd suggest to simply merge Time and Locale preferences into a single window. It could be accessible by both menu entries in Preferences, but each would open it with a different tab activated. This makes sense, because Locale configures how your time is formatted (that is also time-related), and Time configures your timezone (which is clearly alse a Locale thing).

Humdinger said that this would be a lot of tabs, and indeed it would be. However, we should remove the Deskbar-specific settings from there. This is one tab in Time preferences (the one with DeskBar clock settings), and one in Locale preferences (the one labelled with the very undescriptive "Options" and with a single lonely checkbox for "translate application names in Tracker and Deskbar"). I would move these two to the DeskBar settings (that's were these settings for the DeskBar clock used to be, and I already was against moving them back then).
So, it would be only 4 tabs in the end: Time, Timezone, Formatting, Language. Which I think is fine.

For configuring the clock settings, in addition to this DeskBar settings window, I would put them directly in the popup menu of the deskbar clock, so it's easy to change them from there. We can also keep the "Time settings" menu there, which leads to the new window with both Time and Locale formatting settings.

Patch-set: 1