Screen prefs: Layout tweaks based on feedback
Put workspaces settings in their own box below the screen box.

Put brightness slider inside the screen box.

Remove Set background... button, but keep its label for MonitorView
tool tip to let you know that clicking on it opens Backgrounds. There
wasn't enough room for this button with the brightness slider.

When brightness slider is hidden the window is shorter (unless you
have a lot of options on the right.)

Change-Id: I8b39d9e074e7e6abca2c37545f21c456289de381