Update patch set 3
Patch Set 3:

From a logical POV, I think the brightness control belongs into the same group as the screen. The workspaces group could be extra OTOH.
From a visual POV, I think both, the current and the suggested solution look cluttered.

I don't have good ideas on how to improve upon this, though. Maybe get rid of the box frame on the left to save space, and have the brightness label and slider on a single line on that side of the view would look better.

Patch-set: 3
CC: Gerrit User 1000016 <1000016@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>