Update patch set 1
Patch Set 1:

> This is why I'd suggest to simply merge Time and Locale preferences into a single window. It could be accessible by both menu entries in Preferences, but each would open it with a different tab activated.

Yes. But to also get rid of an entry in the Preferences menu of the Deskbar, I suggest to only keep one "Time & Locale" that opens the new combined panel.

>  This is one tab in Time preferences (the one with DeskBar clock settings), and one in Locale preferences (the one labelled with the very undescriptive "Options" and with a single lonely checkbox for "translate application names in Tracker and Deskbar"). I would move these two to the DeskBar settings.

May I suggest not to put the "translate-app-names" into the Deskbar, as it's a system-wide setting. Instead, let's put the lone checkbox to the "use-month/day-names-from-preferred language" checkbox in Locale|Formatting. Not a formatting in the strictest sense, but fits in with names of things (month/day and filenames).

> For configuring the clock settings, in addition to this DeskBar settings window, I would put them directly in the popup menu of the deskbar clock, so it's easy to change them from there. We can also keep the "Time settings" menu there, which leads to the new window with both Time and Locale formatting settings.

Do we need all Clock settings in the Deskbar preferences?
We could rename the "Application" BBox there to "Applications & Tray" and just add a "Hide clock" checkbox there. Then leave the currently 3 Clock settings as checkable items in Clock's context menu.

Patch-set: 1