Update patch set 1
Patch Set 1:

The function uses the hardware timer to measure the CPU clock, apparently. That same timer (and in particular, channel 2) is used to drive the PC speaker at set frequencies. Hearing that activate during the calibration would be unpleasant.

As far as I understandm we use channel 0 on BIOS machines, but that is reserved for EFI own usage, so there we have to use channel 2 still. But we make sure it does not output anything to the speaker during the measurement.

Patch-set: 1