Update patch set 4
Patch Set 4:

Not exactly what I suggested (in case you wanted that in the first place), but it does look less cluttered, even if the price is to remove the backgrounds button.

One could also argue to move the background buttons into the per workspace settings, as the backgrounds can also be per workspace.

Patch-set: 4