Update patch set 1
Patch Set 1:

The problem is not the call to B_LENDIAN_TO_HOST_INT16, but the cast to uint16 from uint8.
On SPARC, the CPU does not handle unaligned memory accesses, and in this case, the uint16 is stored at an odd address (you can check this by looking at the packed structure definition).

So, we have to access it using two separate 8-bit accesses. Yes, the write will probably need to be changed as well, but I did not try to write to disk from the bootloader yet.

We can intercept unaligned accesses (they trigger an hardware exception, pretty much like a segmentation fault) and emulate them in software. This has lower performance but it is probably a good idea anyway, for userland (I don't know if it's reasonable for the kernel, yet). However, for the bootloader it is a bit more complex to setup such handlers, and it seems more reasonable to avoid the unaligned access instead.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>