Tracker: Activate window on pose mouse down
... if we are not in focus follows mouse mode and not dragging.

Disable long press while waiting for drag so that we don't trigger
a context menu to pop up over hovered items during a drag.

Give up on drag after the mouse hasn't moved over a pose for 1/4s.
This gives the user enough time to start a drag comforatably before
the window is activated.

Define kDragSlop in Tracker.h and use that value (3) throughout
Tracker, including in PoseView, to determine the number of pixels
the mouse must move to be in a drag.

Fixes #8824

Change-Id: If3c82d3f0acc4603dc6a8037d399975708436350