Tracker: Activate window on pose mouse up
... unless we are in focus follows mouse mode.

This allows you to start a drag from a background window by clicking
on a pose and dragging it without activating the window. If you wait
too long to start a drag before moving your mouse (double click speed)
you'll trigger a long press which will activate the window.

Don't activate the window on mouse up if we just double clicked on a
selected item because we don't want the Tracker window to activate in
front of the newly opened window.

Activate window on mouse up even when you click on a blank area.

Activate window on drop from a Tracker window.

Fixes #8824

Change-Id: If3c82d3f0acc4603dc6a8037d399975708436350